chore: separate overlay dialog presenter
