+++ src/main/kotlin/com/github/ilovegamecoding/intellijcodexp/presentation/overlay/CodeXPOverlayController.kt
@@ -13,20 +13,7 @@ import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.Service
-import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.wm.WindowManager
-import com.intellij.ui.JBColor
-import java.awt.Color
-import java.awt.Dimension
-import java.awt.Point
-import java.awt.event.MouseAdapter
-import java.awt.event.MouseMotionAdapter
-import javax.swing.BoxLayout
-import javax.swing.JLayeredPane
-import javax.swing.JPanel
-import javax.swing.SwingUtilities
-import javax.swing.Timer
 
 /**
  * Displays CodeXP overlay effects for a single project window.
@@ -43,21 +30,14 @@ class CodeXPOverlayController(
             .getService(CodeXPService::class.java)
 
     private val xpGainEffectRenderer = XpGainEffectRenderer()
-    private val dialogTimers: MutableMap<CodeXPOverlayDialog, Timer> = mutableMapOf()
-    private val dialogDuration: Int = 4000
-    private var ide: JLayeredPane? = null
-    private var dialogArea: JPanel? = null
+    private val dialogPresenter = CodeXPOverlayDialogPresenter(project)
 
     init {
         val connection = ApplicationManager.getApplication().messageBus.connect(this)
         connection.subscribe(CodeXPEventListener.CODEXP_EVENT, this)
         connection.subscribe(CodeXPListener.CODEXP, this)
 
-        SwingUtilities.invokeLater {
-            if (!project.isDisposed) {
-                createDialogArea()
-            }
-        }
+        dialogPresenter.initialize()
     }
 
     override fun eventOccurred(
@@ -83,7 +63,7 @@ class CodeXPOverlayController(
             return
         }
 
-        showDialog(
+        dialogPresenter.show(
             CodeXPOverlayDialog.createDialog(
                 CodeXPBundle.message("overlay.level.up.title"),
                 CodeXPBundle.message(
@@ -114,7 +94,7 @@ class CodeXPOverlayController(
             return
         }
 
-        showDialog(
+        dialogPresenter.show(
             CodeXPOverlayDialog.createDialog(
                 CodeXPBundle.message("overlay.challenge.completed.title"),
                 CodeXPBundle.message("overlay.challenge.completed.main", challenge.name.lowercase()),
@@ -132,100 +112,7 @@ class CodeXPOverlayController(
     }
 
     override fun dispose() {
-        dialogTimers.forEach { (dialog, timer) ->
-            timer.stop()
-            dialog.dispose()
-        }
-        dialogTimers.clear()
-
-        if (SwingUtilities.isEventDispatchThread()) {
-            disposeUi()
-        } else {
-            SwingUtilities.invokeLater {
-                disposeUi()
-            }
-        }
-    }
-
-    private fun disposeUi() {
         xpGainEffectRenderer.dispose()
-        dialogArea?.let { area ->
-            area.removeAll()
-            ide?.remove(area)
-            ide?.revalidate()
-            ide?.repaint()
-        }
-        dialogArea = null
-        ide = null
-    }
-
-    private fun createDialogArea() {
-        val layeredPane =
-            WindowManager
-                .getInstance()
-                .getIdeFrame(project)
-                ?.component
-                ?.rootPane
-                ?.layeredPane
-
-        if (layeredPane == null) {
-            thisLogger().warn("Could not find IDE frame for project ${project.name}.")
-            return
-        }
-
-        val area =
-            JPanel().apply {
-                layout = BoxLayout(this, BoxLayout.Y_AXIS)
-                isVisible = false
-                background = JBColor(Color(255, 255, 255, 0), Color(255, 255, 255, 0))
-                isOpaque = false
-
-                addMouseListener(object : MouseAdapter() {})
-                addMouseMotionListener(object : MouseMotionAdapter() {})
-            }
-
-        ide = layeredPane
-        dialogArea = area
-        layeredPane.add(area, JLayeredPane.POPUP_LAYER, 0)
-    }
-
-    private fun showDialog(dialog: CodeXPOverlayDialog) {
-        val area = dialogArea ?: return
-        val layeredPane = ide ?: return
-
-        dialogTimers[dialog] =
-            Timer(dialogDuration) {
-                hideDialog(dialog)
-            }.apply {
-                isRepeats = false
-                start()
-            }
-        dialog.show()
-
-        with(area) {
-            add(dialog.frame, 0)
-            size = Dimension(480, preferredSize.height)
-            location = Point(layeredPane.width / 2 - width / 2, (layeredPane.height * 0.05).toInt())
-            revalidate()
-            repaint()
-            isVisible = true
-        }
-    }
-
-    private fun hideDialog(dialog: CodeXPOverlayDialog) {
-        SwingUtilities.invokeLater {
-            val area = dialogArea ?: return@invokeLater
-            dialogTimers.remove(dialog)?.stop()
-            dialog.dispose()
-
-            with(area) {
-                remove(dialog.frame)
-                size = Dimension(480, preferredSize.height)
-            }
-
-            if (area.components.isEmpty()) {
-                area.isVisible = false
-            }
-        }
+        dialogPresenter.dispose()
     }
 }
